--- src/fuzzer/runners/python/Types.ts
@@ -1,11 +1,0 @@
-export type PythonEnv = {
-  env: { [k: string]: string | undefined };
-  libs: string | undefined | null;
-  interpreter: string;
-  paths: readonly string[];
-  venv?: {
-    activateCmd: string;
-    path: string;
-    interpreter: string;
-  };
-};
